agent/requirements.txt
flask
flask-cors
langchain
langchain-anthropic
langchain-core
python-dotenv
redis
gunicorn
requests
langfuse
